requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sarvamai
litellm
silero-vad
torch==2.11.0+cpu
numpy
tiktoken
--extra-index-url https://download.pytorch.org/whl/cpu